Added validations to command handlers and fixed middleware

+++ Negotation.Application/Handlers/RemoveProductByIdHandler.cs
@@ -1,5 +1,6 @@
 ﻿using Negotation.Application.Abstractions;
 using Negotation.Application.Commands;
+using Negotation.Application.Exceptions;
 using Negotation.Domain.Repositories;
 
 namespace Negotation.Application.Handlers;
@@ -15,6 +16,13 @@ internal class RemoveProductByIdHandler : ICommandHandler<RemoveProductById>
 
     public async Task HandleAsync(RemoveProductById command)
     {
+        var product = await _repository.GetByIdAsync(command.Id);
+
+        if(product is null)
+        {
+            throw new ProductByIdIsNullException(command.Id.ToString());
+        }
+
         await _repository.RemoveByIdAsync(command.Id);
     }
 }
