+++ Negotation.Application/Handlers/RemoveProductByNameHandler.cs
@@ -1,5 +1,6 @@
 ﻿using Negotation.Application.Abstractions;
 using Negotation.Application.Commands;
+using Negotation.Application.Exceptions;
 using Negotation.Domain.Repositories;
 
 namespace Negotation.Application.Handlers;
@@ -15,6 +16,13 @@ internal class RemoveProductByNameHandler : ICommandHandler<RemoveProductByName>
 
     public async Task HandleAsync(RemoveProductByName command)
     {
+        var product = await _repository.GetByNameAsync(command.Name);
+
+        if (product is null)
+        {
+            throw new ProductByIdNameNullException(command.Name);
+        }
+
         await _repository.RemoveByNameAsync(command.Name);
     }
 }
